requirements.txt
Flask
numpy
opencv-python
Pillow
gTTS
tensorflow
keras
werkzeug
